+++ common/src/main/java/com/railwayteam/railways/content/smokestack/particles/legacy/SmokeParticleData.java
@@ -1,4 +1,4 @@
-package com.railwayteam.railways.content.smokestack;
+package com.railwayteam.railways.content.smokestack.particles.legacy;
 
 import com.mojang.brigadier.StringReader;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;