+++ common/src/main/java/com/railwayteam/railways/mixin_interfaces/IPreAssembleCallback.java
@@ -1,0 +1,5 @@
+package com.railwayteam.railways.mixin_interfaces;
+
+public interface IPreAssembleCallback {
+    void snr$preAssemble();
+}